data augmentation

+++ mlmc/metamodel/flow_dataset.py
@@ -21,8 +21,8 @@ class FlowDataset(Dataset):
     GRAPHS_FILE = "graphs"
     DATA_FILE = "data"
 
-    def __init__(self, output_dir=None, level=0, log=False, mesh=None, corr_field_config=None, config={},
-                 index=None, predict=False, **kwargs):
+    def __init__(self, output_dir=None, level=0, log=False, mesh=None, corr_field_config=None, config={}, index=None,
+                 predict=False, **kwargs):
         self._output_dir = output_dir
         # if self._output_dir is None:
         #     self._output_dir = OUTPUT_DIR
@@ -36,6 +36,7 @@ class FlowDataset(Dataset):
         self._index = index
         self._predict = predict
         self._dataset_config = config.get('dataset_config', {})
+        self._augment_data = config.get('augment_data', False)
 
         if predict:
             self._min_features = self._dataset_config.get('min_features', None)
@@ -58,16 +59,31 @@ class FlowDataset(Dataset):
             self._var_output = None
             self._output_mult_factor = 1
 
+        self._aug_data = []
+        self._columns = None
+
         super().__init__(**kwargs)
 
         #self.a = self.adjacency_matrix
         self.dataset = pd.DataFrame(self.data)
 
+        self._df_for_augmentation = pd.DataFrame(self._aug_data, columns=self._columns)
+
+        #self._data_augmentation()
+
+    def get_train_data(self, index, length):
+        new_dataset = self.dataset[index * length: index * length + length]
+        if self._augment_data:
+            new_dataset = self._data_augmentation(self._df_for_augmentation[index * length: index * length + length])
+
+        new_obj = copy.deepcopy(self)
+        new_obj.dataset = new_dataset
+        return new_obj
+
     def get_test_data(self, index, length):
         new_dataset = self.dataset[0:index * length] + self.dataset[index * length + length:]
         new_obj = copy.deepcopy(self)
         new_obj.dataset = new_dataset
-
         return new_obj
 
     def shuffle(self, seed=None):
@@ -76,6 +92,51 @@ class FlowDataset(Dataset):
 
         random.shuffle(self.data)
         self.dataset = pd.DataFrame(self.data)
+
+    def _data_augmentation(self, df_slice):
+        import smogn
+        import matplotlib.pyplot as plt
+        import seaborn
+
+        # df_slice = self._df_for_augmentation[self._index * self._config['n_train_samples']:
+        #                             self._index * self._config['n_train_samples'] + self._config['n_train_samples']]
+
+        # print("index ", self._index)
+        # print("df slice shape ", df_slice.shape)
+        if 'augmentation_config' in self._config:
+            print("len df slice 0", len(df_slice))
+            dataset_modified = smogn.smoter(data=df_slice, y="y", **self._config["augmentation_config"])
+        else:
+            dataset_modified = smogn.smoter(data=df_slice, y="y", k=9, samp_method="extreme")
+
+        # print("dataset modified shape",  dataset_modified.shape)
+        # print("dataset modified ", dataset_modified)
+        #
+        # print("df stats ", smogn.box_plot_stats(df_slice['y'])['stats'])
+        # print("modified stats ", smogn.box_plot_stats(dataset_modified['y'])['stats'])
+        #
+        # fig, ax = plt.subplots(1, 1, figsize=(15, 10))
+        # ax.hist(df_slice['y'], bins=50, alpha=0.5, label='target', density=True)
+        # plt.title("original")
+        # plt.show()
+        #
+        # fig, ax = plt.subplots(1, 1, figsize=(15, 10))
+        # ax.hist(dataset_modified['y'], bins=50, alpha=0.5, label='target', density=True)
+        # plt.title("modified")
+        # plt.show()
+        #
+        # appended_dataset = df_slice.append(dataset_modified)
+        #
+        # fig, ax = plt.subplots(1, 1, figsize=(15, 10))
+        # ax.hist(appended_dataset['y'], bins=50, alpha=0.5, label='target', density=True)
+        # plt.title("appended")
+        # plt.show()
+
+        return dataset_modified
+
+        # seaborn.kdeplot(df_slice['y'], label="Original")
+        # seaborn.kdeplot(dataset_modified['y'], label="Modified")
+
 
     # def generate_data(self):
     #     n_samples = 10**5
@@ -147,16 +208,19 @@ class FlowDataset(Dataset):
         # max_output = np.max(all_outputs)
 
         if not self._predict:
-            train_outputs = all_outputs[self._index * self._config['n_train_samples']:
-                                        self._index * self._config['n_train_samples'] + self._config['n_train_samples']]
-            train_features = all_features[self._index * self._config['n_train_samples']:
-                                          self._index * self._config['n_train_samples'] + self._config['n_train_samples']]
+            if self._index is not None:
+                train_outputs = all_outputs[self._index * self._config['n_train_samples']:
+                                            self._index * self._config['n_train_samples'] + self._config['n_train_samples']]
+                train_features = all_features[self._index * self._config['n_train_samples']:
+                                              self._index * self._config['n_train_samples'] + self._config['n_train_samples']]
+            else:
+                train_outputs = all_outputs
+                train_features = all_features
 
             if self._dataset_config.get("calc_output_mult_factor", False) is True:
                 self._output_mult_factor = 1/np.mean(train_outputs)
                 self._dataset_config["output_mult_factor"] = self._output_mult_factor
                 self._save_data_config()
-                print("output mult factor ", self._dataset_config["output_mult_factor"])
 
             if self._dataset_config.get("features_normalization", False):
                 self._min_features = np.min(train_features, axis=0)
@@ -219,7 +283,6 @@ class FlowDataset(Dataset):
             if self._log and self._dataset_config.get("features_log", False) is False and self._dataset_config.get("output_log", False) is False:
                 features = np.log(features)
                 output = np.log(output)
-                print("self._log SET")
 
             if self._dataset_config.get("last_log_features", False):
                 features = np.log(features)
@@ -232,6 +295,32 @@ class FlowDataset(Dataset):
             # Save data for pandas dataframe creation, not used with Graph neural network
             self.data.append({'x': features, 'y': output})
 
+            if self._augment_data:
+                if self._columns is None:
+                    d = [pd.DataFrame(features.reshape((features.shape[1], features.shape[0])).tolist()).add_prefix("x_")]
+                    new_df = pd.concat(d, axis=1)
+                    #new_df.insert(loc=0, column="y", value=output)
+                    self._columns = []
+                    for col in new_df.columns:
+                        self._columns.append(col)
+                    self._columns.append("y")
+
+                squeezed_features = list(np.squeeze(features))
+                squeezed_features.append(output)
+                self._aug_data.append(squeezed_features)
+
+
+            #new_df["y"] = output
+            # print("new df y ", new_df["y"])
+            # print("new df .shape ", new_df.shape)
+            #print("new df ", new_df)
+
+            # if self._df_for_augmentation is not None:
+            #     self._df_for_augmentation = self._df_for_augmentation.append(new_df)
+            # else:
+            #     self._df_for_augmentation = new_df
+
+
         self.a = self.adjacency_matrix
         return graphs
 
@@ -240,12 +329,13 @@ class FlowDataset(Dataset):
         import pickle
         import shutil
 
-        if os.path.exists(os.path.join(self._config['iter_dir'], "dataset_config.pkl")):
-            os.remove(os.path.join(self._config['iter_dir'], "dataset_config.pkl"))
+        if "iter_dir" in self._config:
+            if os.path.exists(os.path.join(self._config['iter_dir'], "dataset_config.pkl")):
+                os.remove(os.path.join(self._config['iter_dir'], "dataset_config.pkl"))
 
-        # create a binary pickle file
-        with open(os.path.join(self._config['iter_dir'], "dataset_config.pkl"), "wb") as writer:
-            pickle.dump(self._dataset_config, writer)
+            # create a binary pickle file
+            with open(os.path.join(self._config['iter_dir'], "dataset_config.pkl"), "wb") as writer:
+                pickle.dump(self._dataset_config, writer)
 
     @staticmethod
     def pickle_data(data, output_dir, file_path):
